FIO-9467: rename fixture files
