+++ src/utils/formUtil/fixtures/index.js
@@ -1,5 +1,10 @@
 import tableInsideWizardComp from './tableInsideWizardComponents';
-import components from './components';
-import components2 from './components2';
-import components3 from './components3';
-export { tableInsideWizardComp, components, components2, components3 };
+import textFieldsNestedInLayoutComponents from './textFieldsNestedInLayoutComponents';
+import nestedArrayDataComponents from './nestedArrayDataComponents';
+import mixedLayoutAndDataComponents from './mixedLayoutAndDataComponents';
+export {
+  tableInsideWizardComp,
+  textFieldsNestedInLayoutComponents,
+  nestedArrayDataComponents,
+  mixedLayoutAndDataComponents,
+};
